Add SonarLint folders to .gitignore and PCOO.iml
This commit updates the .gitignore file to ignore the SonarLint and .idea/sonarlint directories that often generate automatic files which are not meant to be version controlled. Additionally, several SonarLint-specific directories have been marked as excluded in the PCOO.iml file as well to keep the repository clean and prevent unwanted files from being committed.